# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/audio/transcriptions": {
      "post": {
        "summary": "音频转录",
        "deprecated": false,
        "description": "将音频转换为文本",
        "operationId": "createTranscription",
        "tags": [
          "OpenAI音频(Audio)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "音频文件",
                    "example": ""
                  },
                  "model": {
                    "type": "string",
                    "example": "whisper-1"
                  },
                  "language": {
                    "type": "string",
                    "description": "ISO-639-1 语言代码",
                    "example": ""
                  },
                  "prompt": {
                    "type": "string",
                    "example": ""
                  },
                  "response_format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "text",
                      "srt",
                      "verbose_json",
                      "vtt"
                    ],
                    "default": "json",
                    "example": "json"
                  },
                  "temperature": {
                    "type": "number",
                    "example": 0
                  },
                  "timestamp_granularities": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "word",
                        "segment"
                      ]
                    },
                    "description": "时间戳粒度；火山极速 ASR 仅支持与 verbose_json 组合",
                    "example": [
                      "segment",
                      "word"
                    ]
                  }
                },
                "required": [
                  "file",
                  "model"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功转录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioTranscriptionResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioTranscriptionResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "segments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "words": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "word": {
                  "type": "string"
                },
                "start": {
                  "type": "number",
                  "description": "开始时间，单位秒"
                },
                "end": {
                  "type": "number",
                  "description": "结束时间，单位秒"
                },
                "confidence": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "Combination": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          }
        ],
        "type": "combination"
      },
      "Combination2": {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}